Moved SPOA's method into corresponding file
